Move useWorkspaces into hooks directory.
