+++ src/analysis/Environments/Environments.ts
@@ -38,7 +38,7 @@ import { contactUsActive } from 'src/libs/state';
 import { elements as styleElements } from 'src/libs/style';
 import { cond, DEFAULT as COND_DEFAULT, formatUSD, makeCompleteDate, withBusyState } from 'src/libs/utils';
 import { GoogleWorkspaceInfo, isGoogleWorkspaceInfo, WorkspaceWrapper } from 'src/libs/workspace-utils';
-import { UseWorkspaces, UseWorkspacesResult } from 'src/workspaces/useWorkspaces.models';
+import { UseWorkspaces, UseWorkspacesResult } from 'src/workspaces/hooks/useWorkspaces.models';
 
 import { DeleteAppModal } from './DeleteAppModal';
 import { DeleteButton } from './DeleteButton';
