+++ src/components/RequesterPaysModal.test.js
@@ -5,8 +5,8 @@ import { renderWithAppContexts as render } from 'src/testing/test-utils';
 
 import RequesterPaysModal from './RequesterPaysModal';
 
-jest.mock('src/workspaces/useWorkspaces', () => ({
-  ...jest.requireActual('src/workspaces/useWorkspaces'),
+jest.mock('src/workspaces/hooks/useWorkspaces', () => ({
+  ...jest.requireActual('src/workspaces/hooks/useWorkspaces'),
   useWorkspaces: jest.fn().mockReturnValue({
     loading: false,
     workspaces: [
