+++ src/pages/workspaces/hooks/useWorkspacesWithSubmissionStats.ts
@@ -5,7 +5,7 @@ import { withErrorReporting } from 'src/libs/error';
 import { useCancellation } from 'src/libs/react-utils';
 import * as Utils from 'src/libs/utils';
 import { WorkspaceSubmissionStats, WorkspaceWrapper as Workspace } from 'src/libs/workspace-utils';
-import { useWorkspaces } from 'src/workspaces/useWorkspaces';
+import { useWorkspaces } from 'src/workspaces/hooks/useWorkspaces';
 
 interface WorkspacesWithSubmissionStatsReturn {
   workspaces: Workspace[];
